+++ extensions/amp-analytics/0.1/test/test-instrumentation.js
@@ -17,12 +17,7 @@
 import {AmpDocSingle} from '../../../../src/service/ampdoc-impl';
 import {
   AnalyticsEventType,
-  ClickEventTracker,
   CustomEventTracker,
-  IniLoadTracker,
-  SignalTracker,
-  TimerEventTracker,
-  VisibilityTracker,
 } from '../events';
 
 import {
@@ -89,157 +84,6 @@ describes.realWin('InstrumentationService', {amp: 1}, env => {
     expect(event.target).to.equal(ampdoc);
     expect(event.type).to.equal('test-event');
     expect(event.vars).to.deep.equal({foo: 'bar'});
-  });
-
-  describe('AnalyticsGroup', () => {
-    let group;
-    let insStub;
-
-    beforeEach(() => {
-      group = service.createAnalyticsGroup(analyticsElement);
-      insStub = sandbox.stub(service, 'addListenerDepr_');
-    });
-
-    it('should create group for the ampdoc root', () => {
-      expect(group.root_).to.equal(root);
-    });
-
-    it('should reject trigger in a disallowed environment', () => {
-      sandbox.stub(root, 'getType').callsFake(() => 'other');
-      allowConsoleError(() => { expect(() => {
-        group.addTrigger({on: 'click', selector: '*'});
-      }).to.throw(/Trigger type "click" is not allowed in the other/); });
-    });
-
-    it('should reject trigger that fails to initialize', () => {
-      sandbox.stub(root, 'getTracker').callsFake(() => {
-        throw new Error('intentional');
-      });
-      expect(() => {
-        group.addTrigger({on: 'click', selector: '*'});
-      }).to.throw(/intentional/);
-    });
-
-    it('should delegate to deprecated addListener', () => {
-      const trackerStub = sandbox.stub(root, 'getTracker');
-      const handler = function() {};
-      group.addTrigger({on: 'scroll'}, handler);
-
-      expect(trackerStub).to.not.be.called;
-      expect(group.listeners_).to.be.empty;
-
-      expect(insStub).to.have.callCount(1);
-
-      expect(insStub.args[0][0].on).to.equal('scroll');
-      expect(insStub.args[0][1]).to.equal(handler);
-      expect(insStub.args[0][2]).to.equal(analyticsElement);
-
-    });
-
-    it('should add "click" trigger', () => {
-      const tracker = root.getTracker('click', ClickEventTracker);
-      const unlisten = function() {};
-      const stub = sandbox.stub(tracker, 'add').callsFake(() => unlisten);
-      const config = {on: 'click', selector: '*'};
-      const handler = function() {};
-      expect(group.listeners_).to.be.empty;
-      group.addTrigger(config, handler);
-      expect(stub).to.be.calledOnce;
-      expect(stub).to.be.calledWith(
-          analyticsElement, 'click', config, handler);
-      expect(group.listeners_).to.have.length(1);
-      expect(group.listeners_[0]).to.equal(unlisten);
-      expect(insStub).to.not.be.called;
-    });
-
-    it('should add "custom" trigger', () => {
-      const tracker = root.getTracker('custom', CustomEventTracker);
-      const unlisten = function() {};
-      const stub = sandbox.stub(tracker, 'add').callsFake(() => unlisten);
-      const config = {on: 'custom-event-1', selector: '*'};
-      const handler = function() {};
-      expect(group.listeners_).to.be.empty;
-      group.addTrigger(config, handler);
-      expect(stub).to.be.calledOnce;
-      expect(stub).to.be.calledWith(
-          analyticsElement, 'custom-event-1', config, handler);
-      expect(group.listeners_).to.have.length(1);
-      expect(group.listeners_[0]).to.equal(unlisten);
-      expect(insStub).to.not.be.called;
-    });
-
-    it('should add "render-start" trigger', () => {
-      const config = {on: 'render-start'};
-      group.addTrigger(config, handler);
-      const tracker = root.getTrackerOptional('render-start');
-      expect(tracker).to.be.instanceOf(SignalTracker);
-
-      const unlisten = function() {};
-      const stub = sandbox.stub(tracker, 'add').callsFake(() => unlisten);
-      const handler = function() {};
-      group.addTrigger(config, handler);
-      expect(stub).to.be.calledOnce;
-      expect(stub).to.be.calledWith(
-          analyticsElement, 'render-start', config, handler);
-    });
-
-    it('should add "ini-load" trigger', () => {
-      const config = {on: 'ini-load'};
-      group.addTrigger(config, handler);
-      const tracker = root.getTrackerOptional('ini-load');
-      expect(tracker).to.be.instanceOf(IniLoadTracker);
-
-      const unlisten = function() {};
-      const stub = sandbox.stub(tracker, 'add').callsFake(() => unlisten);
-      const handler = function() {};
-      group.addTrigger(config, handler);
-      expect(stub).to.be.calledOnce;
-      expect(stub).to.be.calledWith(
-          analyticsElement, 'ini-load', config, handler);
-    });
-
-    it('should add "timer" trigger', () => {
-      const handler = function() {};
-      const unlisten = function() {};
-      const stub = sandbox.stub(TimerEventTracker.prototype, 'add').callsFake(
-          () => unlisten);
-      const config = {on: 'timer'};
-      group.addTrigger(config, handler);
-      const tracker = root.getTrackerOptional('timer');
-      expect(tracker).to.be.instanceOf(TimerEventTracker);
-      expect(stub).to.be.calledOnce;
-      expect(stub).to.be.calledWith(analyticsElement, 'timer', config, handler);
-      expect(group.listeners_).to.have.length(1);
-      expect(group.listeners_[0]).to.equal(unlisten);
-    });
-
-    it('should add "visible" trigger', () => {
-      const config = {on: 'visible'};
-      group.addTrigger(config, handler);
-      const tracker = root.getTrackerOptional('visible');
-      expect(tracker).to.be.instanceOf(VisibilityTracker);
-
-      const unlisten = function() {};
-      const stub = sandbox.stub(tracker, 'add').callsFake(() => unlisten);
-      const handler = function() {};
-      group.addTrigger(config, handler);
-      expect(stub).to.be.calledOnce;
-      expect(stub).to.be.calledWith(
-          analyticsElement, 'visible', config, handler);
-    });
-
-    it('should add "visible" trigger for hidden', () => {
-      group = service.createAnalyticsGroup(analyticsElement);
-      const config = {on: 'hidden'};
-      const getTrackerSpy = sandbox.spy(root, 'getTracker');
-      group.addTrigger(config, () => {});
-      expect(getTrackerSpy).to.be.calledWith('visible');
-      const tracker = root.getTrackerOptional('visible');
-      const unlisten = function() {};
-      const stub = sandbox.stub(tracker, 'add').callsFake(() => unlisten);
-      group.addTrigger(config, () => {});
-      expect(stub).to.be.calledWith(analyticsElement, 'hidden', config);
-    });
   });
 });
